Krazz

+++ app/database/migrations/2015_03_23_121943_create_scheduleinfos_table.php
@@ -19,7 +19,7 @@ class CreateScheduleinfosTable extends Migration {
 			$table->string('end_time');
 			$table->integer('booking_id');
 			$table->integer('day');
-			$table->decimal('price', 10, 10);
+			$table->integer('price');
 			$table->timestamps();
 		});
 	}
